requirements.txt
fastapi[standard]
uvicorn 
SQLAlchemy
psycopg2-binary
pydantic
httpx
pytest
websocket-client

